Split generic part of terminal frontend/terminal/ipapp into core/shellapp
Now there's a class for IPKernel to inherit from in the qt code
